Updated Task2 to filter distinct values
Hi Don,
I looked at the requirements, and we need to calculate the top 20 pages based off unique occurrences of words in the top 5000.  So, the max rank that a page could have is 5000.  It looks like we are currently taking into account all occurrences of words in each page.  I believe that by putting in the 'distinct' filter (line 119) we'll be able to get the correct values for the results.